+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - Filter IUPAC in ingredients for illegal chars
 - Improvements in SDS document generation
 - Ingredients CSV import/export extended to support allergen status and notes impact
-- Ingredients JSON export
+- Ingredients JSON import/export
 - Minor improvements in CSV import
 
 ### Version 6.2