+++ releasenotes.md
@@ -5,4 +5,4 @@ Whats New in v6.3
 - Add additional info fields for suppliers
 - Improvements in SDS document generation - This feature its still in beta state
 - Ingredients CSV import/export extended to support allergen status and notes impact
-- Ingredients JSON export
+- Ingredients JSON import/export